+++ R/describe_missing.R
@@ -2,59 +2,27 @@
 #'
 #' @description Provides a detailed description of missing values in a data frame.
 #' This function reports both absolute and percentage missing values of specified
-#' column lists or scales, following recommended guidelines.
+#' variables.
 #'
-#' @details
-#' In psychology, it is common to ask participants to answer questionnaires in
-#' which people answer several questions about a specific topic. For example,
-#' people could answer 10 different questions about how extroverted they are.
-#' In turn, researchers calculate the average for those 10 questions (called
-#' items). These questionnaires are called (e.g., Likert) "scales" (such as the
-#' Rosenberg Self-Esteem Scale, also known as the RSES).
-#'
-#' Some authors recommend reporting item-level missingness per scale, as well
-#' as a participant's maximum number of missing items by scale. For example,
-#' Parent (2013) writes:
-#'
-#' *I recommend that authors (a) state their tolerance level for missing data by scale
-#' or subscale (e.g., "We calculated means for all subscales on which participants gave
-#' at least 75% complete data") and then (b) report the individual missingness rates
-#' by scale per data point (i.e., the number of missing values out of all data points
-#' on that scale for all participants) and the maximum by participant (e.g., "For Attachment
-#' Anxiety, a total of 4 missing data points out of 100 were observed, with no participant
-#' missing more than a single data point").*
-#'
-#' @param data The data frame to be analyzed.
-#' @param select Variable (or lists of variables) to check for missing values (NAs).
-#' @param scales If you rely on composite scores such as psychological scales
-#' or questionnaires, you can provide the shared suffix among those variables
-#' (as a character vector). This is useful if the variables you want to check
-#' the average of all start with the same name (e.g., `varx`), such as is
-#' commonly the case for Likert scales (such as `varx_1`, `varx_2`, `varx_3`,
-#' etc.).
+#' @inheritParams extract_column_names
+#' @param by Optional character string, indicating the names of one or more
+#' variables in the data frame. If supplied, the data will be split by these
+#' variables and summary statistics will be computed for each group. Useful
+#' for survey data by first reshaping the data to the long format.
+#' @param sort Logical. Whether to sort the result from highest to lowest
+#' percentage of missing data.
 #' @return A dataframe with the following columns:
 #'  - `variable`: Variables selected.
-#'  - `n_columns`: Number of items for selected variables.
-#'  - `n_missing`: Number of missing values for those variables (NA stands for Not
-#'  Available).
-#'  - `n_cells`: Total number of cells (i.e., number of participants multiplied by
-#'  the number of columns, `n_columns`).
-#'  - `missing_percent`: The percentage of missing values (`na` divided by `cells`).
-#'  - `missing_max`: The number of missing values for the participant with the most
-#'  missing values for the selected variables.
-#'  - `missing_max_percent`: The amount of missing values for the participant with
-#'  the most missing values for the selected variables, as a percentage
-#'  (i.e., `missing_max` divided by the number of selected columns, `n_columns`).
-#'  - `all_missing`: The number of participants missing 100% of items for that scale
-#'  (the selected variables).
+#'  - `n_missing`: Number of missing values.
+#'  - `missing_percent`: Percentage of missing values.
+#'  - `complete_percent`: Percentage of non-missing values.
 #' @param ... Arguments passed down to other functions. Currently not used.
 #'
 #' @export
-#' @references Parent, M. C. (2013). Handling item-level missing
-#' data: Simpler is just as good. *The Counseling Psychologist*,
-#' *41*(4), 568-600. https://doi.org/10.1177%2F0011000012445176
 #' @examples
-#' # Use the entire data frame
+#' describe_missing(airquality)
+#'
+#' # Survey data
 #' set.seed(15)
 #' fun <- function() {
 #'   c(sample(c(NA, 1:10), replace = TRUE), NA, NA, NA)
@@ -65,74 +33,78 @@
 #'   extroversion_1 = fun(), extroversion_2 = fun(), extroversion_3 = fun(),
 #'   agreeableness_1 = fun(), agreeableness_2 = fun(), agreeableness_3 = fun()
 #' )
-#' describe_missing(df)
 #'
-#' # If the questionnaire items start with the same name,
-#' # one can list the scale names directly:
-#' describe_missing(df, scales = c("ID", "openness", "extroversion", "agreeableness"))
+#' df_long <- reshape_longer(
+#'   df,
+#'   select = -1,
+#'   names_sep = "_",
+#'   names_to = c("dimension", "item"))
 #'
-#' # Otherwise you can provide nested columns manually:
-#' describe_missing(df,
-#'   select = list(
-#'     c("ID"),
-#'     c("openness_1", "openness_2", "openness_3"),
-#'     c("extroversion_1", "extroversion_2", "extroversion_3"),
-#'     c("agreeableness_1", "agreeableness_2", "agreeableness_3")
-#'   )
-#' )
+#' describe_missing(
+#'   df_long,
+#'   select = -c(1, 3),
+#'   by = "dimension")
 #'
-describe_missing <- function(data, select = NULL, scales = NULL, ...) {
-  vars <- select
-  if (!is.null(vars) && missing(scales)) {
-    vars.internal <- names(data)
-  } else if (!missing(scales)) {
-    vars.internal <- lapply(scales, function(x) {
-      grep(paste0("^", x), names(data), value = TRUE)
+describe_missing <- function(data,
+                             select = NULL,
+                             exclude = NULL,
+                             ignore_case = FALSE,
+                             regex = FALSE,
+                             verbose = TRUE,
+                             by = NULL,
+                             sort = FALSE,
+                             ...) {
+  if (!is.null(select) || !is.null(exclude)) {
+    data <- data_select(
+      data = data,
+      select = select,
+      exclude = exclude,
+      ignore_case = ignore_case,
+      regex = regex,
+      verbose = verbose,
+      ...
+    )
+  }
+  if (!is.null(by)) {
+    if (!by %in% names(data)) {
+      stop("The 'by' column does not exist in the data.")
+    }
+    grouped_data <- split(data, data[[by]])
+    na_list <- lapply(names(grouped_data), function(group_name) {
+      group <- grouped_data[[group_name]]
+      # Identify columns to analyze (exclude the 'by' column)
+      cols_to_analyze <- setdiff(names(group), by)
+      group_na_list <- lapply(cols_to_analyze, function(x) {
+        data_subset <- group[, x, drop = FALSE]
+        .describe_missing(data_subset)
+      })
+      group_na_df <- do.call(rbind, group_na_list)
+      group_na_df$variable <- group_name
+      group_na_df
     })
-  } else if (is.null(vars) && missing(scales)) {
-    vars <- as.list(names(data))
-  }
-  if (!is.null(vars)) {
-    vars.internal <- vars
-  }
-  if (!is.list(vars.internal)) {
-    vars.internal <- list(vars.internal)
-  }
-  na_df <- .describe_missing(data)
-  if (!is.null(vars) || !missing(scales)) {
-    na_list <- lapply(vars.internal, function(x) {
+  } else {
+    na_list <- lapply(names(data), function(x) {
       data_subset <- data[, x, drop = FALSE]
       .describe_missing(data_subset)
     })
-    na_df$variable <- "Total"
-    na_df <- do.call(rbind, c(na_list, list(na_df)))
   }
+  na_df <- do.call(rbind, na_list)
+  if (isTRUE(sort)) {
+    na_df <- na_df[order(-na_df$missing_percent), ]
+  }
+  na_df_tot <- .describe_missing(data)
+  na_df_tot$variable <- "Total"
+  na_df <- rbind(na_df, na_df_tot)
   na_df
 }
 
 .describe_missing <- function(data) {
-  if (ncol(data) > 1) {
-    my_var <- paste0(names(data)[1], ":", names(data)[ncol(data)])
-  } else {
-    my_var <- names(data)
-  }
-  items <- ncol(data)
-  na <- sum(is.na(data))
-  cells <- nrow(data) * ncol(data)
-  na_percent <- round(na / cells * 100, 2)
-  na_max <- max(rowSums(is.na(data)))
-  na_max_percent <- round(na_max / items * 100, 2)
-  all_na <- sum(apply(data, 1, function(x) all(is.na(x))))
-
+  n_missing <- sum(is.na(data))
+  missing_percent <- round(n_missing / (nrow(data) * ncol(data)) * 100, 2)
   data.frame(
-    variable = my_var,
-    n_columns = items,
-    n_missing = na,
-    n_cells = cells,
-    missing_percent = na_percent,
-    complete_percent = 100 - na_percent,
-    missing_max = na_max,
-    missing_max_percent = na_max_percent,
-    all_missing = all_na
+    variable = names(data)[1],
+    n_missing = n_missing,
+    missing_percent = missing_percent,
+    complete_percent = 100 - missing_percent
   )
 }